+++ Tests/Giftcard.py
@@ -3,7 +3,7 @@ import time
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.action_chains import ActionChains
 from selenium.webdriver.support.ui import WebDriverWait
-from Functions import (
+from Utils.Functions import (
     initialize_driver,
     enter_info,
     amenify_one,
