+++ config.json
@@ -1,0 +1,4 @@
+{
+    "email": "aduggal+gcp2@amenify.com",
+    "password": "Akulduggal46@123456"
+}
